# api/openapi.yaml
openapi: 3.1.0
info:
  title: sing-box-panel API
  version: 0.1.0
  license:
    name: GPL-3.0-or-later
    identifier: GPL-3.0-or-later
servers:
  - url: /api/v1
paths:
...
  /sub/{token}:
    servers:
      - url: /
        description: Public endpoint outside the versioned management API. A configured panel base path prefixes this URL at runtime.
    parameters:
      - name: token
        in: path
        required: true
        description: Opaque one-time-issued public subscription credential.
        schema: { type: string, minLength: 1, maxLength: 512 }
    get:
      operationId: getPublicSubscription
      security: []
      description: Returns only bytes frozen into the currently applied activation bundle. Token revocation and expiry are checked immediately; mutable channels and sources are never read for publication.
      parameters:
        - name: format
          in: query
          description: Required for an unbound token. Optional for a channel-bound token, where it must match the bound channel.
          schema: { $ref: "#/components/schemas/SubscriptionFormat" }
      responses:
        "200":
          description: Frozen client subscription bytes from the applied bundle.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
            application/yaml:
              schema: { type: string }
            text/plain:
              schema: { type: string }
        "400":
          description: The explicit format or query string is invalid.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: Uniform response for an unknown, expired, revoked, mismatched, or unavailable token/channel selection.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "503":
          description: No bundle has been applied, or frozen publication is temporarily unavailable. The response never reflects the token.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /subscription/channels:
    get:
      operationId: listSubscriptionChannels
      responses:
        "200":
          description: All subscription channels in stable name order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SubscriptionChannel" }
        "401": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    post:
      operationId: createSubscriptionChannel
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionChannelWrite" }
      responses:
        "201":
          description: The channel was created. ETag is its updated_at CAS version.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionChannel" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /subscription/channels/{channelId}:
    parameters:
      - $ref: "#/components/parameters/SubscriptionChannelID"
    get:
      operationId: getSubscriptionChannel
      responses:
        "200":
          description: One channel. ETag is its updated_at CAS version.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionChannel" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    put:
      operationId: updateSubscriptionChannel
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/SubscriptionIfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionChannelWrite" }
      responses:
        "200":
          description: The channel after the CAS update.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionChannel" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "428": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteSubscriptionChannel
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/SubscriptionIfMatch"
      responses:
        "204": { description: The unreferenced channel was deleted. }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "428": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /subscription/channels/{channelId}/preview:
    parameters:
      - $ref: "#/components/parameters/SubscriptionChannelID"
    post:
      operationId: previewSubscriptionChannel
      description: Renders from one ready or stale immutable startup artifact without changing public publication.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionPreviewRequest" }
      responses:
        "200":
          description: Renderer result and stable diagnostics.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionPreview" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /subscription/sources:
    get:
      operationId: listSubscriptionSources
      responses:
        "200":
          description: All attached sources in stable name order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SubscriptionSource" }
        "401": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
    post:
      operationId: createSubscriptionSource
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionSourceCreate" }
      responses:
        "201":
          description: The source was created. ETag is its updated_at CAS version.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionSource" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
  /subscription/sources/{sourceId}/snapshot:
    parameters:
      - $ref: "#/components/parameters/SubscriptionSourceID"
    put:
      operationId: updateSubscriptionSourceSnapshot
      description: Stores a new strict source snapshot candidate. Public publication changes only after a later successful apply freezes it.
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/SubscriptionIfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionSourceSnapshotWrite" }
      responses:
        "200":
          description: The source after the snapshot CAS update.
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionSource" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "428": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
  /subscription/tokens/{tokenId}:
    parameters:
      - $ref: "#/components/parameters/SubscriptionTokenID"
    get:
      operationId: getSubscriptionToken
      description: Returns metadata only; never plaintext or the persisted digest.
      responses:
        "200":
          description: Token metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionToken" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "500": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    managementBearer:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: sbp_session
  parameters:
...
    SubscriptionChannelID:
      name: channelId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
...
  responses:
    Problem:
      description: Stable RFC 9457-compatible problem response.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        request_id: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
...
    SubscriptionFormat:
      type: string
      enum: [sing-box, mihomo, loon]
...
    SubscriptionChannel:
      type: object
      additionalProperties: false
      required: [id, name, format, config, enabled, created_at, updated_at]
      properties:
        id: { type: string, minLength: 1, maxLength: 128 }
        name: { type: string, minLength: 1, maxLength: 128 }
        format: { $ref: "#/components/schemas/SubscriptionFormat" }
        config: { $ref: "#/components/schemas/SubscriptionChannelConfig" }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
security:
  - sessionCookie: []
  - managementBearer: []